remove instrumentation part, leave only default test entry point for android
